move redeploy params logic to its own class
