+++ app/helpers/deploys_helper.rb
@@ -79,7 +79,7 @@ module DeploysHelper
       project_stage_deploys_path(
         @project,
         @deploy.stage,
-        deploy: redeploy_params
+        deploy: Samson::RedeployParams.new(@deploy).to_hash
       ),
       html_options
   end
@@ -105,42 +105,5 @@ module DeploysHelper
       end
 
     path_to_image(favicon)
-  end
-
-  private
-
-  # Applies different logic depending on the class of each of the deploy
-  # parameters, so it supports nested paramaters based on object relations
-  def redeploy_params
-    params = Samson::Hooks.fire(:deploy_permitted_params).flatten(1)
-    params.each_with_object(reference: @deploy.reference) do |param, collection|
-      case param
-      when String, Symbol
-        collection[param] = @deploy.public_send(param)
-      when Hash
-        nested_redeploy_params(collection, param)
-      else
-        raise "Unsupported deploy param class: `#{param.class}` for `#{param}`."
-      end
-    end
-  end
-
-  def nested_redeploy_params(collection, params)
-    params.each_with_object(collection) do |(key, attributes), nested|
-      if key.to_s.end_with?('attributes')
-        nested[key] = deploy_relation_attributes(key, attributes)
-      end
-    end
-  end
-
-  # currently, this only supports `has_many` relations, cause the public_method
-  # we call on the instance is expected to return an array
-  def deploy_relation_attributes(key, attributes)
-    relation_name = key.to_s.sub(/_attributes$/, '')
-    @deploy.public_send(relation_name).map do |item|
-      (attributes - [:id, :_destroy]).each_with_object({}) do |attribute, hash|
-        hash[attribute] = item.public_send(attribute)
-      end
-    end
   end
 end
